YAY!! Notifications work (mostly) right!
So when you tap a notification, it uses the MessageListActivity that's on the
top of the stack. Which is ideal, as it doesn't convolute the hierarchy
of the app.

BUT if the app is not running and you tap a notification, it launches a
single MessageListActivity that is not based on a ThreadListActivity.
This means that the back button exits the app. :(

Working on this.