Change style. Perfect notifications.
Now, the thread you are in hangs around between multitasking (before it was
destroyed when you stopped the activity) AND when a notification comes, it
gets re-initialized using onNewIntent to become the other thread. Which is
awesome. Yay!

TODO: test things like starting as one thread, switching/becoming another via
a notification, and then using the add/leave/etc.. functions to make sure
they still act on the proper thread.